LUA/TEX: Define Committee faction
This is my silliest project yet. A new faction based around a stock market and buying access to units from each of the other factions. Whether or not any of this will work is a huge question mark.

Note that this also reverts 28b574. I wasn't going to wait 24 hours to branch and commit this stuff, but I also don't want Stonks on it.